:recycle: NavBar 컴포넌트 업데이트
