--- frontend/src/components/NavBar/NavBar.tsx
@@ -1,7 +1,0 @@
-import * as S from './NavBar.css';
-
-const NavBar = () => {
-  return <nav className={S.Layout}>NavBar</nav>;
-};
-
-export default NavBar;
