+++ frontend/src/pages/Layout/Layout.tsx
@@ -1,6 +1,6 @@
 import { Outlet } from 'react-router-dom';
 import Header from '@/components/_common/Header/Header';
-import NavBar from '@/components/NavBar/NavBar';
+import NavBar from '@/components/_common/NavBar/NavBar';
 
 const Layout = () => {
   return (
